backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
bcrypt
pyjwt
python-multipart
email-validator
